feat: sqlite
orms have to be like the worst thing to ever come out in computing. i
spent just 5 minutes looking at the seaorm docs and just felt immensely
repulsed. i've been evangelicized by the pure church of SQL. raw SQL is
how god intended for man to interact with the database, nothing more and
nothing less. the syntax is simple and matches the behavior directly.
raw SQL invokes the same feelings as controlling your allocations
directly when writing. we are not a slop shop so we should not be
pushing out slop. we are craftspeople building beautiful tools and we
must be precise and direct with what we write.

just as the humble carpenter knows every grain in the slabs of wood they
choose, we must know every single query that exits the pyrobot, an
robotic amalgamation of all of us who have ever touched this repository.
the orm abstraction is like attempting to sculpt man out of marble and
putting on welding gloves to avoid the chisel from slipping and hitting
your hand. we must understand every last behavior of our program, to
understand exactly what is happening and when it is happening for when
god made man, we were made to sculpt and to craft.

for man to give up control over its very creation is comparable to the
farmer letting wheat take dominion of his farm and grow itself at its
own whim. we must treat every line of code as an
addition to our holy temple. would you walk through our temple with
muddy boot on? no, so why would you introduce the unholy orm, reaper of
sin? through the humble SQL query, we are able to speak directly and
clearly to the database, unimpeded by the once unknown horrors of object
oriented design. each query a humble and selfless prayer, pure and
uncorrupted, directly appealing to the great powers.

look upon the code soiled by the orm and despair. for ye who reaps sin
will also sow the seeds of distraught and suffering that follow. to veil
the truth of SQL is to commit a great cardinal sin, one that will
forever stain your purity and transitively the performance of every line
of code you write.

seaorm is cursed by the mark of the beast. listen to this final appeal
and submit to the final and eternal truth of sqlx.